update gradle plugin helper package path
