+++ GoRouter-Gradle-Plugin/src/main/kotlin/com/wyjson/router/gradle_plugin/core/doc/GenerateRouteDocTask.kt
@@ -1,12 +1,11 @@
-package com.wyjson.router.gradle_plugin.helper
+package com.wyjson.router.gradle_plugin.core.doc
 
 import com.android.build.api.variant.Variant
 import com.google.gson.Gson
-import com.wyjson.router.gradle_plugin.helper.model.RouteHelperModel
+import com.google.gson.GsonBuilder
+import com.wyjson.router.gradle_plugin.core.doc.model.DocumentModel
 import com.wyjson.router.gradle_plugin.utils.Constants
-import com.wyjson.router.gradle_plugin.utils.Constants.GOROUTER_HELPER_CLASS_NAME
-import com.wyjson.router.gradle_plugin.utils.Constants.HELPER_TAG
-import com.wyjson.router.gradle_plugin.utils.Constants.NO_FOUND_HELPER_TAG_PLUGIN_TIPS
+import com.wyjson.router.gradle_plugin.utils.Constants.GENERATE_ROUTE_DOC
 import com.wyjson.router.gradle_plugin.utils.Logger
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
@@ -16,7 +15,7 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
 import java.io.File
 
-abstract class GenerateGoRouterHelperTask : DefaultTask() {
+abstract class GenerateRouteDocTask : DefaultTask() {
 
     init {
         group = Constants.PROJECT
@@ -25,52 +24,40 @@ abstract class GenerateGoRouterHelperTask : DefaultTask() {
     @get:Input
     abstract var variant: Variant
 
-    private val TAG = "RH"
+    private val TAG = "Doc"
 
     private var dependModeList: ArrayList<String> = ArrayList()
-    private var routeHelperModel: RouteHelperModel? = null
-
-    private val catalog: String = "main" // main or variantName or buildType
-    private var rootProject: Project? = null
+    private var document: DocumentModel? = null
 
     @TaskAction
     fun taskAction() {
-        Logger.i(TAG, "Generate GoRouterHelper task start.")
+        Logger.i(TAG, "Generate GoRouter document task start.")
         val variantName = variant.name
         val buildType = variant.buildType
         val flavorName = variant.flavorName
         setDependModeList(variantName, buildType, flavorName)
 
         if (!scanRouteModule(variantName, buildType)) return
-        if (rootProject == null){
-            Logger.e(TAG, NO_FOUND_HELPER_TAG_PLUGIN_TIPS.trimIndent())
-            return
-        }
+        val json = GsonBuilder().setPrettyPrinting().create().toJson(document)
 
-        val className = GOROUTER_HELPER_CLASS_NAME
-        val dir = rootProject!!.buildDir
-        val path = "/generated/source/gorouter/${catalog}/com/wyjson/router/${className}.java"
-        val outputFile = File(dir, path)
-        outputFile.parentFile.mkdirs()
-        outputFile.writeText(AssembleGoRouteHelperCode(routeHelperModel!!).toJavaCode(className), Charsets.UTF_8)
-        Logger.i(TAG, "Generate GoRouterHelper task end. ${dir}${path}")
+        val documentName = "${project.rootProject.name}-${variantName}-${Constants.DOCUMENT_FILE_NAME}"
+        val docOutFilePath = "${project.projectDir}/${documentName}"
+        File(docOutFilePath).writeText(json, Charsets.UTF_8)
+        Logger.i(TAG, "Generate GoRouter document[${documentName}] task end. $docOutFilePath")
     }
 
     private fun scanRouteModule(variantName: String, buildType: String?): Boolean {
         project.dependProject().plus(project).forEach { curProject ->
-            if (curProject.isHelperTag()) {
-                rootProject = curProject
-            }
             var file = searchJSONFile(curProject, variantName)
             if (file == null && buildType != null) {
-                 file = searchJSONFile(curProject, buildType)
+                file = searchJSONFile(curProject, buildType)
             }
             if (file != null) {
-                mergeRouteModule(curProject, file)
+                mergeRouteModuleDoc(curProject, file)
             }
         }
-        if (routeHelperModel == null) {
-            Logger.e(TAG, "Failed to generate the route!")
+        if (document == null) {
+            Logger.e(TAG, "Failed to generate the route document!")
             return false
         }
         return true
@@ -84,31 +71,32 @@ abstract class GenerateGoRouterHelperTask : DefaultTask() {
             collection = curProject.files(genFile).asFileTree.filter { it.name.endsWith(Constants.DOCUMENT_FILE_NAME) }
         }
         if (collection.isEmpty) {
-            Logger.w(TAG, "project[${curProject.name}] scan 0 route.")
+            Logger.w(TAG, "project[${curProject.name}] scan 0 route document.")
             return null
         } else {
             val file = collection.first()
             Logger.i(TAG, "project[${curProject.name}] found the file[${file.name}].")
-//            catalog = name
             return file
         }
     }
 
-    private fun mergeRouteModule(curProject: Project, file: File) {
+    private fun mergeRouteModuleDoc(curProject: Project, file: File) {
         if (file.readLines().isNotEmpty()) {
             try {
-                val model = Gson().fromJson(file.readText(), RouteHelperModel::class.java)
-                if (routeHelperModel == null) {
-                    routeHelperModel = model
+                val documentModel = Gson().fromJson(file.readText(), DocumentModel::class.java)
+                if (document == null) {
+                    document = documentModel
                 } else {
-                    routeHelperModel!!.services.putAll(model.services)
-                    routeHelperModel!!.routes.putAll(model.routes)
+                    document!!.services.putAll(documentModel.services)
+                    document!!.interceptors.addAll(documentModel.interceptors)
+                    document!!.routes.putAll(documentModel.routes)
                 }
+                document!!.interceptors.sortBy { interceptor -> interceptor.ordinal }
             } catch (e: Exception) {
-                Logger.e(TAG, "module[${curProject.name}] route json parsing failed, do not modify the generated route json, use the '${Constants.GENERATE_ROUTE_DOC}' task to generate a new route json.")
+                Logger.e(TAG, "module[${curProject.name}] route document parsing failed, do not modify the generated route file, use the '${GENERATE_ROUTE_DOC}' task to generate a new route document.")
             }
         } else {
-            Logger.e(TAG, "module[${curProject.name}] route json content is empty and a new route json is generated using the '${Constants.GENERATE_ROUTE_DOC}' task.")
+            Logger.e(TAG, "module[${curProject.name}] route document content is empty and a new route document is generated using the '${GENERATE_ROUTE_DOC}' task.")
         }
     }
 
@@ -150,7 +138,5 @@ abstract class GenerateGoRouterHelperTask : DefaultTask() {
     }
 
     private fun Project.isAndroid() = plugins.hasPlugin("com.android.application") || plugins.hasPlugin("com.android.library")
-    private fun Project.isHelperTag() = plugins.hasPlugin(HELPER_TAG)
-
 
 }
