+++ GoRouter-Gradle-Plugin/src/main/kotlin/com/wyjson/router/gradle_plugin/core/doc/model/DocumentModel.kt
@@ -1,4 +1,4 @@
-package com.wyjson.router.gradle_plugin.doc.model
+package com.wyjson.router.gradle_plugin.core.doc.model
 
 data class DocumentModel(
     val services: HashMap<String, ServiceModel> = HashMap(),
