+++ GoRouter-Gradle-Plugin/src/main/kotlin/com/wyjson/router/gradle_plugin/helper/core/GenerateGoRouterHelperTask.kt
@@ -1,11 +1,12 @@
-package com.wyjson.router.gradle_plugin.doc
+package com.wyjson.router.gradle_plugin.helper.core
 
 import com.android.build.api.variant.Variant
 import com.google.gson.Gson
-import com.google.gson.GsonBuilder
-import com.wyjson.router.gradle_plugin.doc.model.DocumentModel
+import com.wyjson.router.gradle_plugin.helper.model.RouteHelperModel
 import com.wyjson.router.gradle_plugin.utils.Constants
-import com.wyjson.router.gradle_plugin.utils.Constants.GENERATE_ROUTE_DOC
+import com.wyjson.router.gradle_plugin.utils.Constants.GOROUTER_HELPER_CLASS_NAME
+import com.wyjson.router.gradle_plugin.utils.Constants.HELPER_TAG_PACKAGE_NAME
+import com.wyjson.router.gradle_plugin.utils.Constants.NO_FOUND_HELPER_TAG_PACKAGE_NAME_PLUGIN_TIPS
 import com.wyjson.router.gradle_plugin.utils.Logger
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
@@ -15,7 +16,7 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
 import java.io.File
 
-abstract class GenerateRouteDocTask : DefaultTask() {
+abstract class GenerateGoRouterHelperTask : DefaultTask() {
 
     init {
         group = Constants.PROJECT
@@ -24,40 +25,52 @@ abstract class GenerateRouteDocTask : DefaultTask() {
     @get:Input
     abstract var variant: Variant
 
-    private val TAG = "Doc"
+    private val TAG = "RH"
 
     private var dependModeList: ArrayList<String> = ArrayList()
-    private var document: DocumentModel? = null
+    private var routeHelperModel: RouteHelperModel? = null
+
+    private val catalog: String = "main" // main or variantName or buildType
+    private var rootProject: Project? = null
 
     @TaskAction
     fun taskAction() {
-        Logger.i(TAG, "Generate GoRouter document task start.")
+        Logger.i(TAG, "Generate GoRouterHelper task start.")
         val variantName = variant.name
         val buildType = variant.buildType
         val flavorName = variant.flavorName
         setDependModeList(variantName, buildType, flavorName)
 
         if (!scanRouteModule(variantName, buildType)) return
-        val json = GsonBuilder().setPrettyPrinting().create().toJson(document)
+        if (rootProject == null){
+            Logger.e(TAG, NO_FOUND_HELPER_TAG_PACKAGE_NAME_PLUGIN_TIPS.trimIndent())
+            return
+        }
 
-        val documentName = "${project.rootProject.name}-${variantName}-${Constants.DOCUMENT_FILE_NAME}"
-        val docOutFilePath = "${project.projectDir}/${documentName}"
-        File(docOutFilePath).writeText(json, Charsets.UTF_8)
-        Logger.i(TAG, "Generate GoRouter document[${documentName}] task end. $docOutFilePath")
+        val className = GOROUTER_HELPER_CLASS_NAME
+        val dir = rootProject!!.buildDir
+        val path = "/generated/source/gorouter/${catalog}/com/wyjson/router/${className}.java"
+        val outputFile = File(dir, path)
+        outputFile.parentFile.mkdirs()
+        outputFile.writeText(AssembleGoRouteHelperCode(routeHelperModel!!).toJavaCode(className), Charsets.UTF_8)
+        Logger.i(TAG, "Generate GoRouterHelper task end. ${dir}${path}")
     }
 
     private fun scanRouteModule(variantName: String, buildType: String?): Boolean {
         project.dependProject().plus(project).forEach { curProject ->
+            if (curProject.isHelperTag()) {
+                rootProject = curProject
+            }
             var file = searchJSONFile(curProject, variantName)
             if (file == null && buildType != null) {
-                file = searchJSONFile(curProject, buildType)
+                 file = searchJSONFile(curProject, buildType)
             }
             if (file != null) {
-                mergeRouteModuleDoc(curProject, file)
+                mergeRouteModule(curProject, file)
             }
         }
-        if (document == null) {
-            Logger.e(TAG, "Failed to generate the route document!")
+        if (routeHelperModel == null) {
+            Logger.e(TAG, "Failed to generate the route!")
             return false
         }
         return true
@@ -71,32 +84,31 @@ abstract class GenerateRouteDocTask : DefaultTask() {
             collection = curProject.files(genFile).asFileTree.filter { it.name.endsWith(Constants.DOCUMENT_FILE_NAME) }
         }
         if (collection.isEmpty) {
-            Logger.w(TAG, "project[${curProject.name}] scan 0 route document.")
+            Logger.w(TAG, "project[${curProject.name}] scan 0 route.")
             return null
         } else {
             val file = collection.first()
             Logger.i(TAG, "project[${curProject.name}] found the file[${file.name}].")
+//            catalog = name
             return file
         }
     }
 
-    private fun mergeRouteModuleDoc(curProject: Project, file: File) {
+    private fun mergeRouteModule(curProject: Project, file: File) {
         if (file.readLines().isNotEmpty()) {
             try {
-                val documentModel = Gson().fromJson(file.readText(), DocumentModel::class.java)
-                if (document == null) {
-                    document = documentModel
+                val model = Gson().fromJson(file.readText(), RouteHelperModel::class.java)
+                if (routeHelperModel == null) {
+                    routeHelperModel = model
                 } else {
-                    document!!.services.putAll(documentModel.services)
-                    document!!.interceptors.addAll(documentModel.interceptors)
-                    document!!.routes.putAll(documentModel.routes)
+                    routeHelperModel!!.services.putAll(model.services)
+                    routeHelperModel!!.routes.putAll(model.routes)
                 }
-                document!!.interceptors.sortBy { interceptor -> interceptor.ordinal }
             } catch (e: Exception) {
-                Logger.e(TAG, "module[${curProject.name}] route document parsing failed, do not modify the generated route file, use the '${GENERATE_ROUTE_DOC}' task to generate a new route document.")
+                Logger.e(TAG, "module[${curProject.name}] route json parsing failed, do not modify the generated route json, use the '${Constants.GENERATE_ROUTE_DOC}' task to generate a new route json.")
             }
         } else {
-            Logger.e(TAG, "module[${curProject.name}] route document content is empty and a new route document is generated using the '${GENERATE_ROUTE_DOC}' task.")
+            Logger.e(TAG, "module[${curProject.name}] route json content is empty and a new route json is generated using the '${Constants.GENERATE_ROUTE_DOC}' task.")
         }
     }
 
@@ -138,5 +150,7 @@ abstract class GenerateRouteDocTask : DefaultTask() {
     }
 
     private fun Project.isAndroid() = plugins.hasPlugin("com.android.application") || plugins.hasPlugin("com.android.library")
+    private fun Project.isHelperTag() = plugins.hasPlugin(HELPER_TAG_PACKAGE_NAME)
+
 
 }
