+++ GoRouter-Gradle-Plugin/src/main/resources/META-INF/gradle-plugins/com.wyjson.gorouter.helper.properties
@@ -1,0 +1,1 @@
+implementation-class=com.wyjson.router.gradle_plugin.helper.launch.HelperGradlePluginLaunch
