+++ GoRouter-Gradle-Plugin/src/main/resources/META-INF/gradle-plugins/com.wyjson.gorouter.helper.tag.properties
@@ -1,1 +1,1 @@
-implementation-class=com.wyjson.router.gradle_plugin.launch.HelperTagGradlePluginLaunch
+implementation-class=com.wyjson.router.gradle_plugin.helper.launch.HelperTagGradlePluginLaunch
