+++ GoRouter-Gradle-Plugin/src/main/resources/META-INF/gradle-plugins/com.wyjson.gorouter.properties
@@ -1,1 +1,1 @@
-implementation-class=com.wyjson.router.gradle_plugin.launch.GradlePluginLaunch
+implementation-class=com.wyjson.router.gradle_plugin.launch.MainGradlePluginLaunch
