+++ app/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.application'
     id 'com.wyjson.gorouter'
+    id 'com.wyjson.gorouter.helper'
 }
 GoRouter {
     // 允许执行自动注册任务的集合,最好不要写debug,以节省开发阶段build时间
